refactor: ExceptionHandler 위치 변경
