+++ src/main/java/parser/InputParser.java
@@ -12,6 +12,4 @@ public class InputParser {
         return LocalTime.of(hour,minute);
     }
 
-
-
 }
