+++ src/main/java/util/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import static constant.Command.QUIT;
